requirements.txt
# Core Framework
phidata==2.4.0
typer==0.12.0
rich==13.7.0
python-dotenv==1.0.0

# LLM & Embeddings
openai==1.12.0
anthropic==0.18.1

# RAG & Vector Stores
langchain==0.1.10
langchain-community==0.0.25
sentence-transformers==2.5.1
chromadb==0.4.24
rank-bm25==0.2.2

# Database
psycopg[binary]==3.1.18
psycopg2-binary==2.9.9
sqlalchemy==2.0.27
pgvector==0.2.5
redis==5.0.1

# Market Data
yfinance==0.2.36

# SEC Filing Parsing
beautifulsoup4==4.12.3
lxml==5.1.0
pdfplumber==0.11.0
pypdf==4.0.1
requests==2.31.0

# Utilities
pandas==2.2.1
numpy==1.26.4
python-dateutil==2.8.2

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-mock==3.12.0

# Observability
structlog==24.1.0
prometheus-client==0.20.0

# Development
black==24.2.0
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.2